requirements.txt
langchain
langchain-openai
langgraph
fastapi
uvicorn
python-dotenv
langchain-community